+++ src/mod/internal/rail_mod_base.cpp
@@ -69,7 +69,7 @@ void RailModBase::rdp_input_invalidate(Rect r)
     }
 }
 
-void RailModBase::rdp_input_mouse(int device_flags, int x, int y)
+void RailModBase::rdp_input_mouse(uint16_t device_flags, uint16_t x, uint16_t y)
 {
     if (device_flags & (MOUSE_FLAG_WHEEL | MOUSE_FLAG_HWHEEL)) {
         x = this->old_mouse_x;
