+++ src/mod/internal/replay_mod.hpp
@@ -57,7 +57,7 @@ public:
 
     void rdp_input_invalidate(Rect /*rect*/) override {}
 
-    void rdp_input_mouse(int /*device_flags*/, int /*x*/, int /*y*/) override {}
+    void rdp_input_mouse(uint16_t /*device_flags*/, uint16_t /*x*/, uint16_t /*y*/) override {}
 
     void rdp_input_scancode(KbdFlags flags, Scancode scancode, uint32_t event_time, Keymap const& keymap) override;
