+++ src/mod/internal/widget/label.hpp
@@ -80,7 +80,7 @@ public:
 
     void set_color(Color bg_color, Color fg_color) override;
 
-    void rdp_input_mouse(int device_flags, int x, int y) override;
+    void rdp_input_mouse(uint16_t device_flags, uint16_t x, uint16_t y) override;
 
     void auto_resize();
