+++ src/mod/internal/widget/module_host.cpp
@@ -611,7 +611,7 @@ void WidgetModuleHost::rdp_input_invalidate(Rect clip)
     }
 }
 
-void WidgetModuleHost::rdp_input_mouse(int device_flags, int x, int y)
+void WidgetModuleHost::rdp_input_mouse(uint16_t device_flags, uint16_t x, uint16_t y)
 {
     this->current_pointer_pos_x = x;
     this->current_pointer_pos_y = y;
