+++ src/mod/internal/widget/module_host.hpp
@@ -107,7 +107,7 @@ public:
 
     void rdp_input_invalidate(Rect clip) override;
 
-    void rdp_input_mouse(int device_flags, int x, int y) override;
+    void rdp_input_mouse(uint16_t device_flags, uint16_t x, uint16_t y) override;
 
     void rdp_input_scancode(KbdFlags flags, Scancode scancode, uint32_t event_time, Keymap const& keymap) override;
 
@@ -151,6 +151,6 @@ private:
 
     gdi::CachePointerIndex current_cache_pointer_index;
 
-    int current_pointer_pos_x = 0;
-    int current_pointer_pos_y = 0;
+    uint16_t current_pointer_pos_x = 0;
+    uint16_t current_pointer_pos_y = 0;
 };
